attached lregister function to form component
